v1.1
grid of nodes now created. after converting the grid code from c# unity to c++ with sfml.
Next to implement click features to set types nodes(start and end) and starting to implementing a*. Ususally wouldnt need to draw but to test need to draw which took most of the time.